W-15854291: Keeping the FileSystemProvider, FtpAbstractConnectionProvider and FtpConnectionProvider hierarchy same as previous since its required by FTPS Connector.
